Refactor team filter
