+++ foosballv2s/Source/Activities/Adapters/TeamAdapter.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Android.Content;
 using Android.Views;
 using Android.Widget;
+using foosballv2s.Filters;
 
 namespace foosballv2s.Adapters
 {
@@ -56,47 +57,6 @@ namespace foosballv2s.Adapters
         public override Filter Filter {
             get {
                 return filter;
-            }
-        }
-
-        class TeamFilter : Filter
-        {
-            TeamAdapter teamAdapter;
-            
-            public TeamFilter (TeamAdapter adapter) : base() {
-                teamAdapter = adapter;
-            }
-            protected override Filter.FilterResults PerformFiltering (Java.Lang.ICharSequence constraint)
-            {
-                FilterResults results = new FilterResults();
-                if (constraint != null) {
-
-                    if (teamAdapter.IgnoreFilter)
-                    {
-                        teamAdapter.IgnoreFilter = false;
-                    }
-                    else
-                    {
-                       teamAdapter.SelectedTeam = null; 
-                    }
-                    
-                    var searchFor = constraint.ToString();
-                    var matchList = new List<Team>();
-                    
-                    var matches = from i in teamAdapter.teams
-                        where i.TeamName.Contains(searchFor)
-                        select i;
-
-                    teamAdapter.matchTeams = matches.ToList<Team>();
-                    
-//                    results.Values = matchObjects;
-                    results.Count = matchList.Count;
-                }
-                return results;
-            }
-            protected override void PublishResults (Java.Lang.ICharSequence constraint, Filter.FilterResults results)
-            {
-                teamAdapter.NotifyDataSetChanged();
             }
         }
     }
